StatementConfiguration: factor out of StatementContext
Fixes #626
