+++ core/src/main/java/org/jdbi/v3/core/statement/StatementConfiguration.java
@@ -1,0 +1,102 @@
+package org.jdbi.v3.core.statement;
+
+import java.util.Arrays;
+
+import org.jdbi.v3.core.config.JdbiConfig;
+
+/**
+ * Contains configuration applied to Statements created by the {@code DefaultStatementBuilder}
+ * or your own {@link StatementBuilder} implementation.
+ */
+public class StatementConfiguration implements JdbiConfig<StatementConfiguration>
+{
+    private boolean  returningGeneratedKeys;
+    private String[] generatedKeysColumnNames;
+    private boolean  concurrentUpdatable;
+
+    StatementConfiguration() { }
+
+    StatementConfiguration(StatementConfiguration other)
+    {
+        this.returningGeneratedKeys = other.returningGeneratedKeys;
+        this.generatedKeysColumnNames = other.generatedKeysColumnNames.clone();
+        this.concurrentUpdatable = other.concurrentUpdatable;
+    }
+
+    @Override
+    public StatementConfiguration createCopy()
+    {
+        return new StatementConfiguration(this);
+    }
+
+    /**
+     * Control returning generated keys on created statements.
+     * @param b return generated keys?
+     */
+    public void setReturningGeneratedKeys(boolean b)
+    {
+        if (isConcurrentUpdatable() && b) {
+            throw new IllegalArgumentException("Cannot create a result set that is concurrent "
+                    + "updatable and is returning generated keys.");
+        }
+        this.returningGeneratedKeys = b;
+    }
+
+    /**
+     * @return whether the statement being generated is expected to return generated keys.
+     */
+    public boolean isReturningGeneratedKeys()
+    {
+        return returningGeneratedKeys || generatedKeysColumnNames != null && generatedKeysColumnNames.length > 0;
+    }
+
+    /**
+     * @return the generated key column names, if any
+     */
+    public String[] getGeneratedKeysColumnNames()
+    {
+        if (generatedKeysColumnNames == null) {
+            return new String[0];
+        }
+        return Arrays.copyOf(generatedKeysColumnNames, generatedKeysColumnNames.length);
+    }
+
+    /**
+     * Set the generated key column names.
+     * @param generatedKeysColumnNames the generated key column names
+     */
+    public void setGeneratedKeysColumnNames(String[] generatedKeysColumnNames)
+    {
+        this.generatedKeysColumnNames = Arrays.copyOf(generatedKeysColumnNames, generatedKeysColumnNames.length);
+    }
+
+    /**
+     * Return if the statement should be concurrent updatable.
+     *
+     * If this returns true, the concurrency level of the created ResultSet will be
+     * {@link java.sql.ResultSet#CONCUR_UPDATABLE}, otherwise the result set is not updatable,
+     * and will have concurrency level {@link java.sql.ResultSet#CONCUR_READ_ONLY}.
+     *
+     * @return if the statement generated should be concurrent updatable.
+     */
+    public boolean isConcurrentUpdatable() {
+        return concurrentUpdatable;
+    }
+
+    /**
+     * Set the context to create a concurrent updatable result set.
+     *
+     * This cannot be combined with {@link #isReturningGeneratedKeys()}, only
+     * one option may be selected. It does not make sense to combine these either, as one
+     * applies to queries, and the other applies to updates.
+     *
+     * @param concurrentUpdatable if the result set should be concurrent updatable.
+     */
+    public void setConcurrentUpdatable(final boolean concurrentUpdatable) {
+        if (concurrentUpdatable && isReturningGeneratedKeys()) {
+            throw new IllegalArgumentException("Cannot create a result set that is concurrent "
+                    + "updatable and is returning generated keys.");
+        }
+        this.concurrentUpdatable = concurrentUpdatable;
+    }
+}
