+++ core/src/main/java/org/jdbi/v3/core/statement/StatementContext.java
@@ -23,7 +23,6 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collections;
 import java.util.LinkedHashSet;
 import java.util.List;
@@ -47,10 +46,10 @@ import org.jdbi.v3.core.mapper.ColumnMapper;
 import org.jdbi.v3.core.mapper.RowMapper;
 
 /**
- * The statement context provides a means for passing client specific information through the
- * evaluation of a statement. The context is not used by jDBI internally, but will be passed
- * to all statement customizers. This makes it possible to parameterize the processing of
- * the tweakable parts of the statement processing cycle.
+ * The statement context provides access to statement-local configuration.
+ * Context is inherited from the parent {@link Handle} initially and changes will
+ * not outlive the statement.
+ * The context will be passed through most major jdbi APIs.
  *
  * DISCLAIMER: The class is not intended to be extended. The final modifier is absent to allow
  * mock tools to create a mock object of this class in the user code.
@@ -67,9 +66,6 @@ public class StatementContext implements Closeable
     private PreparedStatement statement;
     private Connection        connection;
     private Binding           binding = new Binding();
-    private boolean           returningGeneratedKeys;
-    private boolean           concurrentUpdatable;
-    private String[]          generatedKeysColumnNames;
 
     StatementContext() {
         this(new ConfigRegistry());
@@ -362,65 +358,5 @@ public class StatementContext implements Closeable
     public ExtensionMethod getExtensionMethod()
     {
         return extensionMethod;
-    }
-
-    public void setReturningGeneratedKeys(boolean b)
-    {
-        if (isConcurrentUpdatable() && b) {
-            throw new IllegalArgumentException("Cannot create a result set that is concurrent "
-                    + "updatable and is returning generated keys.");
-        }
-        this.returningGeneratedKeys = b;
-    }
-
-    /**
-     * @return whether the statement being generated is expected to return generated keys.
-     */
-    public boolean isReturningGeneratedKeys()
-    {
-        return returningGeneratedKeys || generatedKeysColumnNames != null && generatedKeysColumnNames.length > 0;
-    }
-
-    public String[] getGeneratedKeysColumnNames()
-    {
-        if (generatedKeysColumnNames == null) {
-            return new String[0];
-        }
-        return Arrays.copyOf(generatedKeysColumnNames, generatedKeysColumnNames.length);
-    }
-
-    public void setGeneratedKeysColumnNames(String[] generatedKeysColumnNames)
-    {
-        this.generatedKeysColumnNames = Arrays.copyOf(generatedKeysColumnNames, generatedKeysColumnNames.length);
-    }
-
-    /**
-     * Return if the statement should be concurrent updatable.
-     *
-     * If this returns true, the concurrency level of the created ResultSet will be
-     * {@link java.sql.ResultSet#CONCUR_UPDATABLE}, otherwise the result set is not updatable,
-     * and will have concurrency level {@link java.sql.ResultSet#CONCUR_READ_ONLY}.
-     *
-     * @return if the statement generated should be concurrent updatable.
-     */
-    public boolean isConcurrentUpdatable() {
-        return concurrentUpdatable;
-    }
-
-    /**
-     * Set the context to create a concurrent updatable result set.
-     *
-     * This cannot be combined with {@link #isReturningGeneratedKeys()}, only
-     * one option may be selected. It does not make sense to combine these either, as one
-     * applies to queries, and the other applies to updates.
-     *
-     * @param concurrentUpdatable if the result set should be concurrent updatable.
-     */
-    public void setConcurrentUpdatable(final boolean concurrentUpdatable) {
-        if (concurrentUpdatable && isReturningGeneratedKeys()) {
-            throw new IllegalArgumentException("Cannot create a result set that is concurrent "
-                    + "updatable and is returning generated keys.");
-        }
-        this.concurrentUpdatable = concurrentUpdatable;
     }
 }
